scripts/requirements.txt
anthropic>=0.40.0
requests>=2.32.0
python-dotenv>=1.0.1
pytest>=8.0.0
